Add the rest of the buttons for temp homepage
